Updated some entads with new info from "Undone"
The wand, bident, coin, and barnacle sword were identified by Filera in ch70.
(The dagger, blanket, and goggles were also identified, but not much new was learned about those. By implication, I've marked them as unbound.)
Also, I changed the name of Hannah's armor (just for fun) and realized that the Oeyr plate thing must be unbound if they are giving it out on loan.
Also changed location of wardrobe based on ch68.